+++ distill/commands/claude_md.py
@@ -17,7 +17,7 @@ import typer
 
 from distill._console import console
 from distill.commands import _logic
-from distill.commands._logic import _complete_topics
+from distill.commands._helpers import _complete_topics
 from distill.library import claude_md
 
 __all__ = ["claude_md_cmd", "register"]